production/golf-enrichment/requirements.txt
# FastAPI and server dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP client for API calls
httpx>=0.26.0

# Async runtime (must be >=4.6 for mcp dependency compatibility)
anyio>=4.6.0

# Claude Agent SDK (core dependency)
# This will install: anyio>=4.0.0, typing_extensions>=4.0.0, mcp>=0.1.0
claude-agent-sdk>=0.1.3

# Supabase client (for Agent 8)
supabase>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Optional: Monitoring and logging
python-json-logger>=2.0.7
